+++ src/pages/course-info/components/CourseTable.tsx
@@ -2,7 +2,7 @@ import { useMemo, useState, useCallback, useRef } from "react";
 import styles from "./course-table.module.css";
 import { FaPen, FaCalendarAlt, FaPlus, FaMinus } from "react-icons/fa";
 import type { Row, Test } from "../types/course-table.types";
-import { makeTests, onlyDigits,formatPhone,validateDraft } from "../../../utils/helper-functions/courseTableHelpers"
+import { makeTests, onlyDigits,formatPhone,validateDraft } from "../../../utils/helpers/courseTableHelpers"
 
 
 export default function CourseTable() {
