+++ src/utils/helpers/calendar-helpers.ts
@@ -37,3 +37,16 @@ export const toMinutes = (hhmm: string) => {
   const [h, m] = hhmm.split(":").map(Number);
   return h * 60 + m;
 };
+
+export const COLOR_PALETTE = [
+  "#E6E8FA", "#EFD3F5", "#FAD6E7", "#FFE0F2",
+  "#D1F7FF", "#C8F7E4", "#E4F9D4", "#FFF2B3",
+  "#FFE7A7", "#FBE0A1", "#FFD8B5", "#FFD6A5",
+  "#CDE1FF", "#BFE8FF", "#BDE0FE", "#D7F3FF",
+  "#D0F0C0", "#C2F0D6", "#F0FFC2", "#FFF7CC",
+  "#F1E0FF", "#E0D1FF", "#E8F0FF", "#F0F4FF",
+];
+
+export const HEBREW_DAY_NAMES = ["ראשון", "שני", "שלישי", "רביעי", "חמישי", "שישי", "שבת"];
+
+export const MAX_DESCRIPTION_CHARS = 300;
