+++ src/utils/helpers/courseTableHelpers.ts
